So much stuff
-Vec3f is now Vec3d and vec3d is also Vec3d
-Tree of weapons/attachments works pretty much properly
-Something's still off with AbsolutePosition when aiming up and to the
side though
-Giant laser cluster diamond thing
-Guns have a fire rate
-Structs automatically generate attach points based on normals: need to
be careful to not generate stuff inside other stuff
-Convenience methods for adding different shapes to Model
-New ModelConstants: basicStrutModel, longStrutModel (temp until random
generation)
-Lighting is not applied to lasers
-Fiddled with the camera, but commented it out

+++ Animation.h
@@ -11,7 +11,7 @@
 
 #include <iostream>
 
-class Vec3f;
+class Vec3d;
 class Model;
 class GameObject;
